WIP fix: set state in constructor, to contain it in instance
@patriciagarcia that was the only problem with your code :) Problem was that you defined
state outside the constructor function, so it remained the same for all tests, and in
the second test, state.pouchDBChangesFeedEmitter was already defined and attached
to the previous api instance, in which you did not add farther docs. That is the reason
why the test with 2 docs passed when run alone, but failed when run with the rest.

Makes sense?